major refactor and speedup
