--- nodes/sharp/models/encoders/__init__.py
@@ -1,24 +1,0 @@
-"""Contains different encoders for Gaussian predictor.
-
-For licensing see accompanying LICENSE file.
-Copyright (C) 2025 Apple Inc. All Rights Reserved.
-"""
-
-from .base_encoder import BaseEncoder
-
-from .monodepth_encoder import (
-    MonodepthFeatureEncoder,
-    create_monodepth_encoder,
-)
-from .spn_encoder import SlidingPyramidNetwork
-from .unet_encoder import UNetEncoder
-from .vit_encoder import create_vit
-
-__all__ = [
-    "create_vit",
-    "BaseEncoder",
-    "UNetEncoder",
-    "SlidingPyramidNetwork",
-    "MonodepthFeatureEncoder",
-    "create_monodepth_encoder",
-]
