--- nodes/sharp/utils/__init__.py
@@ -1,5 +1,0 @@
-"""Contains utils packages.
-
-For licensing see accompanying LICENSE file.
-Copyright (C) 2025 Apple Inc. All Rights Reserved.
-"""
